requirements.txt
streamlit
pandas
openai
python-dotenv
pytest
pytest-cov
numpy
airtable
